[QuickAccent] check_excluded_app_with_title function moved to excluded_apps.h
